+++ flink-table/flink-table-planner/src/test/java/org/apache/flink/table/planner/plan/nodes/exec/stream/AsyncCalcRestoreTest.java
@@ -37,6 +37,7 @@ public class AsyncCalcRestoreTest extends RestoreTestBase {
                 AsyncCalcTestPrograms.ASYNC_CALC_UDF_SIMPLE,
                 AsyncCalcTestPrograms.ASYNC_CALC_UDF_COMPLEX,
                 AsyncCalcTestPrograms.ASYNC_CALC_UDF_NESTED,
-                AsyncCalcTestPrograms.ASYNC_CALC_UDF_CONDITION);
+                AsyncCalcTestPrograms.ASYNC_CALC_UDF_CONDITION,
+                AsyncCalcTestPrograms.ASYNC_CALC_UDF_FAILURE_EXCEPTION);
     }
 }
